THRIFT-5219: Binding capabilities for client socket
Client: cpp

In some telco grade applications binding the source address when using a
client socket is a must to ensure strong trust relationship in some
client/server applications.

This patch has been included to provide an overloaded constructor for
TSocket which enables to define explicitly an IP address to bind to
before connecting the socket to a remote host.